+++ bcompiler/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "1.3.3_experimental"
+__version__ = "1.3.3"