+++ docs/changes.rst
@@ -1,6 +1,14 @@
 Changes
 ~~~~~~~
 
+v1.3.3
+######
+
+* new swimlane assurance milestones analyser
+* annex analyser now does comparison with previous master document
+* fix issues in annex analyser
+*  
+
 v1.3.2
 ######